requirements.txt
sklearn
pandas
torchvision